% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addFile.R
\name{addFile}
\alias{addFile}
\title{add a new code book, data file, document or R file to project}
\usage{
addFile(file.name, move = TRUE, projectDir = "myProject", replace = FALSE)
}
\arguments{
\item{file.name}{name of file to add to project as a string}

\item{move}{whether to move instead of copy file Default: TRUE}

\item{projectDir}{base directory of \code{dryworkflow} project
directory. Default: \dQuote{myProject}}

\item{replace}{logical, if file exists then replace. Default: FALSE}
}
\value{
logical TRUE/FALSE indicating success of adding file
}
\description{
add a code book, data or document file to a \code{dryworkflow}
project. If appropriate, relevant R files will be created and
Makefiles, log files and git repository updated
}

