% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backupFiles.R
\name{backupFiles}
\alias{backupFiles}
\title{Backup original, R files, R report and posted files to allow regenerating}
\usage{
backupFiles(zipFile = NULL)
}
\arguments{
\item{zipFile}{string containing name for zip file. Default:
\code{NULL} whereby name is derived from today's date and the
project directory name}
}
\value{
None
}
\description{
Zips all relevant files to allow work to be reproduced although,
for safety, it is best to make your own backup.
}
\details{
While \code{backupFiles} provides a way to backup all relevant
files to enable work to be reproduced, it is best to have a more
comprehensive personalised strategy tailored to your own
circumstances. In particular, it may be more prudent to modify
\code{Makefile} to backup off site by using make. See Makefile and
type \code{make help rsync} at the command or shell prompt to help
automate remote backup. Note that this function should be called
from the main project directory not a work/reporting sub directory
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}

