% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copyCommonMk.R
\name{copyCommonMk}
\alias{copyCommonMk}
\title{Copy dryworkflow package common.mk file to specified destination}
\usage{
copyCommonMk(destination = NULL, overwriteFile = FALSE, createDir = FALSE)
}
\arguments{
\item{destination}{string containing directory name for copying
\code{common.mk}. Default: "~/lib" for unix style set ups and
$HOME/Library for windows style set ups}

\item{overwriteFile}{logical indicating whether to overwrite
existing \code{common.mk} file: Default: FALSE}

\item{createDir}{whether to create destination directory if it
doesn't exist: Default = FALSE}
}
\value{
None
}
\description{
The file \code{common.mk} contains pattern rules to process
\code{R}, \code{Rmd} and \code{Rnw} files to output a range of
output files including \code{Rout}, \code{pdf}, \code{html} and
\code{docx} files
}
\details{
The \code{common.mk} file is supplied with the \code{dryworkflow}
package. Ideally, this file should be placed in a specific
directory used for all data analysis projects. In \code{linux}
this would usually be something like \code{~/lib}. The definitions
and rules can then be used for any project by including the
contents with an \code{include} command placed near the end of a
\code{Makefile}. Individual definitions or rules can be overridden
by redefining them after the \code{include} statement.  The latest
version of \code{common.mk} is always available at
\url{https://github.com/petebaker/r-makefile-definitions}. Once a
basic Makefile is set up (usually by
\code{\link{createProjectSkeleton}}) then type \code{make help}
for more details.
}
\examples{
copyCommonMk("testMake", createDir = TRUE)
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}

