% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copyTemplates.R
\name{copyTemplates}
\alias{copyTemplates}
\title{Copy dryworkflow package template files to specified destination}
\usage{
copyTemplates(destination = "templates")
}
\arguments{
\item{destination}{string containing directory name for copying
template files. Default: a new directory "templates" in the
current directory}
}
\value{
None
}
\description{
Copies all template \code{.txt} files from package
\code{dryworkflow} to a directory so that they can be modified and
reused. Files can be customised and then used by
\code{\link{createProjectSkeleton}} by specifying an alternative
\code{template.dir}. Note that if the directory specified already
exists then an error will be produced.
}
\details{
Files have names like \code{template_cleanR.txt},
\code{template_readR.txt}, \code{template_analyseR.txt} and so
on. Their use should be obvious from the file name. Strings which
may be changed are described at the top of the template file and
the description will be removed from the syntax file which is
produced using \code{\link{createProjectSkeleton}}.
}
\examples{
copyTemplates()
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}

