% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createFromTemplate.R
\name{createFromTemplate}
\alias{createFromTemplate}
\title{Create an R, Rmd or Rnw file from a template}
\usage{
createFromTemplate(file.name, subst.strings, template, template.dir,
  overwrite.file = FALSE, string.prefix = "@@",
  string.suffix = string.prefix,
  delete.start = "-- START: DELETE THIS SECTION --",
  delete.end = "-- END: DELETE THIS SECTION --", print.mismatches = FALSE)
}
\arguments{
\item{file.name}{full filename of file to be written provided as a
string and including directory if necessary}

\item{subst.strings}{named list of string substitutions}

\item{template}{name of template text file as string. Default:
\code{NULL} for predefined template from \code{\link{dryworkflow}}
package.}

\item{template.dir}{directory containing template. Default:
\code{\link{dryworkflow}} package directory}

\item{overwrite.file}{logical whether or not to allow overwrite of
existing file. Default: \code{FALSE}}

\item{string.prefix}{string of characters for start of strings to
be substituted from template file. Default \sQuote{@@}}

\item{string.suffix}{string of characters for end of strings to be
substituted from template file. Default: same as
\code{string.prefix}}

\item{delete.start}{lines between and including those containing
the \code{delete.start} and \code{delete.end} patterns will be
removed. Default: \dQuote{---- START: DELETE THIS SECTION ----}}

\item{delete.end}{Default: \dQuote{---- END: DELETE THIS SECTION ----}}

\item{print.mismatches}{logical to declare wther to print warnings
about unused and undefined strings. Default: \code{FALSE}}
}
\value{
None
}
\description{
Creates and writes an \code{R} syntax or report file from a
template by substituting for specified strings. By default, these
templates are provided as part of the \code{\link{dryworkflow}}
package. In order to customise these files to a particular project
or style of working, alternate templates and string formats can be
provided.
}
\details{
By default, strings in the template file that look like
@@SYNTAX.FILE@@ and @@DATA.FILE@@ are substituted with
strings provided as elements in a list with named components
SYNTAX.FILE, DATA.FILE and so on provided as the \code{subst.strings}
argument. The string prefix and suffix can be changed but must be
the same throughout the template.
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}

