% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createSyntaxR.R
\name{createSyntaxR}
\alias{createSyntaxR}
\title{Create \code{R} syntax for reading, cleaning, summarising and analysing}
\usage{
createSyntaxR(dir.project, filesAndDFs, template, project.steps = c("read",
  "clean", "summary", "analyse", "mergeAll", "codebook", "reportRmd",
  "reportRnw", "presentRmd", "beamerRmd", "beamerRnw"),
  makefile.depends = NULL, makefile.targets = c("Rout", "pdf"),
  myFunction.files = NULL, libraries = NULL, template.dir = NULL,
  print.mismatches = FALSE, overwrite.file = FALSE, string.prefix = "@@",
  string.suffix = string.prefix,
  delete.start = "-- START: DELETE THIS SECTION --",
  delete.end = "-- END: DELETE THIS SECTION --", extras = list(LICENCE =
  options()$dryworkflow$licence, AUTHOR = options()$dryworkflow$author,
  TITLE.PROJECT = options()$dryworkflow$title.project))
}
\arguments{
\item{dir.project}{directory name for project}

\item{filesAndDFs}{object of S3 class \code{fileAndDataName}
containing relevant filenames, directories for setting up syntax
and Makefiles}

\item{template}{name of template text file as string. Default:
\code{NULL} for predefined template from \code{\link{dryworkflow}}
package.}

\item{project.steps}{steps to be carried out in project, specified
as a vector of strings. Options are \dQuote{read} to read data
(always assumed present), \dQuote{clean} clean data,
\dQuote{summary} summary statistics and basic plots,
\dQuote{analyse} perform statistical analysis, \dQuote{compare}
compare datasets and in particular different versions of the same
data set, \dQuote{mergeAll} merge data sets of more than one; and
\dQuote{reportRmd} or \dQuote{reportRnw} produce reports using
\code{\link{rmarkdown}} and/or\code{\link{Sweave}} and
\dQuote{presentRmd} or \dQuote{beamerRnw} produce presentations
using \code{\link{rmarkdown}} and/or\code{\link{Sweave}}}

\item{makefile.depends}{files to be used as dependencies in
addition to the syntax file for targets. Default: worked out from
project structure (\code{NULL})}

\item{makefile.targets}{strings with file extensions for targets
in makefiles Default: c(\dQuote{Rout}, \dQuote{pdf}) using
\code{R BATCH} and \code{stitch} via \code{rmarkdown}, respectively}

\item{myFunction.files}{character vector of own homegrown function
file names to be sourced not currently put in to a package}

\item{libraries}{character vector of library statements to be
added near top of \code{R} syntax file. Default: load
\code{dryworkflow} \code{Hmisc}, \code{foreign}}

\item{template.dir}{directory containing template. Default:
\code{\link{dryworkflow}} package directory}

\item{print.mismatches}{print mismatches when creating syntax or
markdown files from templates. Default: \code{FALSE}}

\item{overwrite.file}{logical whether or not to allow overwrite of
existing file. Default: FALSE}

\item{string.prefix}{string of characters for start of strings to
be substituted from template file. Default \sQuote{@@}}

\item{string.suffix}{string of characters for end of strings to be
substituted from template file. Default: same as
\code{string.prefix}}

\item{delete.start}{lines between and including those containing
the \code{delete.start} and \code{delete.end} patterns will be
removed. Default: \dQuote{---- START: DELETE THIS SECTION ----}}

\item{delete.end}{Default: \dQuote{---- END: DELETE THIS SECTION ----}}

\item{extras}{list of extra options to be passed to function for
substitution in template. Default: licence, author and
title.project obtained from global options}

\item{AUTHOR}{string containing Author's name for \code{R} and
markdown syntax files}

\item{TITLE.PROJECT}{string containing title of project for log
files and reports}
}
\value{
Lines to be included in Makefile for reading files as
character vector
}
\description{
Function is used internally by \code{\link{createProjectSkeleton}} and
\code{\link{addFile}} to produce individual \code{.R} template
syntax files for cleaning, summarising and analysing data files in
a data analysis project
}
\author{
Peter Baker  \email{pete@petebaker.id.au}
}

