% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{extractSubstStrings}
\alias{extractSubstStrings}
\title{extract strings for substitution from template for checking etc}
\usage{
extractSubstStrings(template, string.prefix = "@@",
  string.suffix = string.prefix)
}
\arguments{
\item{template}{character vector where each element is the lines
of a template file}

\item{string.prefix}{string of characters for start of strings to
be substituted. Default \sQuote{@@}}

\item{string.suffix}{string of characters for end of strings to be
substituted. Default: same as \code{string.prefix}}
}
\value{
vector of strings for substitution
}
\description{
Given a template vector \code{template} read in from a template
file and string format via \code{string.prefix} and
\code{string.sufffix}, extract a vector of all strings to be
substituted.
}
\seealso{
\code{\link{readTemplate}} reads a template file into a
string, \code{\link{createFromTemplate}} creates a new file from a template
}
\keyword{internal}

