% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProjectConfig.R
\name{getProjectConfig}
\alias{getProjectConfig}
\title{Get dryworkflow configuration from configFile.rds file}
\usage{
getProjectConfig(projectDir = ".", checkAbove = FALSE,
  checkSubDirs = FALSE)
}
\arguments{
\item{projectDir}{base directory of dryWorkflow project}

\item{checkAbove}{check directory above current directory for
presence of configuration file}

\item{checkSubDirs}{check directories below current directory for
presence of configuration file(s). If there is more than one then
print locations. If there is only one, print a message indicating
which project directory is present and use that.}
}
\value{
object of class \dQuote{drywProjectConfig} else FALSE if
file \code{configFile.rds} is not found or if object not of
correct class
}
\description{
Configuration file \code{configFile.rds} is created with
\code{\link{createProjectSkeleton}} when a project is created. It
contains details like directory structures and various parameters
for a project.
}

