% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readCodeBook.R
\name{readCodeBook}
\alias{readCodeBook}
\title{Read a code book in standard format as a csv file}
\usage{
readCodeBook(x, codebook.directory = NULL, col.names = c(var.name =
  "Variable", var.orig = "Original.Name", var.label = "Label", levels =
  "Levels", min = "Min", max = "Max"), non.standard = NULL,
  na.strings = c("", "NA", ".", " "))
}
\arguments{
\item{x}{filename of codebook to parse}

\item{codebook.directory}{directory containing codebook}

\item{col.names}{named character vector containing column names in
Code Book file. The vector contains components \dQuote{var.name} =
variable name, \dQuote{var.orig} = original name (if changed),
\dQuote{label} for printing/plotting, \dQuote{level} = factor
levels, \dQuote{min} and \dQuote{max} for continuous measurements,
\dQuote{comments} = comments about the variable which may include
the measurement instrument or references about the measurement}

\item{non.standard}{named list of non-standard names of columns
with names \code{c("var.name", "var.orig", "var.label", "levels", "min", "max")}}

\item{na.strings}{a character vector of strings which are to be
interpreted as \sQuote{NA} values.  Blank fields are also
considered to be missing values in logical, integer, numeric and
complex fields. Default: \code{c("", "NA", ".", " ")}}
}
\value{
S3 object of type class \dQuote{codebook}
}
\description{
Reads a code book stored as a \code{csv} file for either checking
against a data file or relabelling factor levels or labelling
variables.
}
\details{
Often, data dictionaries or code books are provided with data
files. Rather than a \code{word} \code{doc} or \code{pdf} files,
the format required here is in a very specific format stored as a
\code{csv} file. Once read in, attributes such as factor
labels/levels and variable labels can be added to the
\code{data.frame} and/or also used to check factor labels and
variable names are consistent with the code book. Note that while
various methods may be available which attempt to convert word
docs or pdf's to a spreadsheet and/or csv file, extreme care
should be taken as these are far from perfect.
}
\examples{
file.copy(system.file('demoFiles', 'data1_codebook.csv',
                      package='dryworkflow'), 'data1_codebook.csv')
data1_codebook <- readCodeBook("data1_codebook.csv",
         non.standard = list(levels = "Factor.Levels",
                             var.orig = "Old.Variable"))
}
\author{
Peter Baker  \email{pete@petebaker.id.au}
}

