% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{readTemplate}
\alias{readTemplate}
\title{Read template file into a string}
\usage{
readTemplate(template, template.dir = NULL,
  delete.start = "-- START: DELETE THIS SECTION --",
  delete.end = "-- END: DELETE THIS SECTION --")
}
\arguments{
\item{template}{name of template text file as string. Default:
\code{NULL} for predefined template from \code{\link{dryworkflow}}
package.}

\item{template.dir}{directory containing template. Default:
\code{\link{dryworkflow}} package directory}

\item{delete.start}{lines between and including those containing
the \code{delete.start} and \code{delete.end} patterns will be
removed. Default: \dQuote{---- START: DELETE THIS SECTION ----}}

\item{delete.end}{Default: \dQuote{---- END: DELETE THIS SECTION
----}}
}
\value{
vector string containing complete \code{template} with a
string for each line of the template file
}
\description{
Internal function used by functions when checking or creating
syntax and markdown files
}
\details{
Note that lines between (and including) lines containing the
patterns \dQuote{---- START: DELETE THIS SECTION ----} and
\dQuote{---- END: DELETE THIS SECTION ----} will be deleted from
the template
}
\author{
Peter Baker \email{pete@petebaker.id.au}
}
\keyword{internal}

