% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setUpDirectoryStructure.R
\name{setUpDirectoryStructure}
\alias{setUpDirectoryStructure}
\title{Set up directories for data analysis project}
\usage{
setUpDirectoryStructure(style = NULL, type.project = NULL,
  destinations = NULL, extra = NULL, extra.destinations = NULL, ...)
}
\arguments{
\item{style}{style for directory and file names (\code{unix} or
\code{windows}), Default: \code{unix}}

\item{type.project}{type of project: normal, simple or custom
(custom Not Yet Implemented). The style of directory structure
for the project. Default: \dQuote{normal}}

\item{destinations}{\code{list} of destination directories where
original (and added) files will be moved. This is a \code{list}
with named components \sQuote{data}, \sQuote{doc},
\sQuote{codebook} and \sQuote{lib} for data files, documents,
codebooks and R functions, respectively. Default: unix directory
names will be \code{list(data = "data/original", doc = "doc/original", codebook = "data/codebook", lib = "lib")} and
Windows will be of similar form.}

\item{extra}{extra directories additional to standard setup}

\item{extra.destinations}{extra destination directories additional
to standard setup}

\item{...}{extra arguments passed to specific functions}
}
\value{
\code{list} of directories including destinations for
initial files of class \sQuote{drywDestinationDirs}. Named
components are \code{directories} and \code{destinations}.
}
\description{
Set up directories for data analysis project
}

