% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupFileAndDataNames.R
\name{setupFileAndDataNames}
\alias{setupFileAndDataNames}
\title{Internal: Create syntax filenames and object names for processing}
\usage{
setupFileAndDataNames(dir.project, destinations, projectConfig, projectDirs,
  filenames = NULL, initial.files = NULL, mnemonic = "",
  project.steps = c("read", "codebook", "clean", "summary", "analyse",
  "compare", "mergeAll", "reportRmd", "reportRnw", "presentRmd", "beamerRmd",
  "beamerRnw"), report.markdown = c("Rmd", "Rnw"), report.which = c("first",
  "merge", "all"))
}
\arguments{
\item{dir.project}{dir.project directory name for project. Default:
\dQuote{myRproject}}

\item{destinations}{\code{list} of destination directories where
original (and added) files will be moved. This is a \code{list}
with named components \sQuote{data}, \sQuote{doc},
\sQuote{codebook} and \sQuote{lib} for data files, documents,
codebooks and R functions, respectively. Default: unix directory
names will be \code{list(data = "data/original", doc =
"doc/original", codebook = "data/codebook", lib = "lib")} and
Windows will be of similar form but capitalised
with longer form names.}

\item{projectConfig}{project configuration stored at project
creation and updated when files added. Format similar to similar
to getOptions(\dQuote{dryworkflow})}

\item{projectDirs}{directory structure of project of class
'drywDestinationDirs'}

\item{filenames}{filenames for added files but not set for new
project.}

\item{initial.files}{initial file sources and extensions}

\item{mnemonic}{three or four letter mnemonic to aid remembering
and for succinct naming \code{R}, \code{Rmd} and \code{Rnw} files
and project directory. Default: \code{NULL} for none}

\item{project.steps}{steps to be carried out in project, specified
as a vector of strings. Options are \dQuote{read} to read data
(always assumed present), \dQuote{clean} clean data,
\dQuote{summary} summary statistics and basic plots,
\dQuote{analyse} perform statistical analysis, \dQuote{compare}
compare datasets and in particular different versions of the same
data set,  \dQuote{mergeAll} merge data sets of more than one; and
\dQuote{reportRmd} or \dQuote{reportRnw} produce reports using
\code{\link{rmarkdown}} and/or\code{\link{Sweave}} and
\dQuote{presentRmd} or \dQuote{beamerRnw} produce presentations
using \code{\link{rmarkdown}} and/or\code{\link{Sweave}}}

\item{report.markdown}{vector of markdown file types to be
employed to produce reports such as \dQuote{.org}, \dQuote{.Rmd}
and \dQuote{.Rnw}. Default: \dQuote{.Rmd} and \dQuote{.Rnw}.}

\item{report.which}{which data files to produce reports
for. Choices: \dQuote{first}, \dQuote{merge}, \dQuote{all})
Default: \dQuote{first}}
}
\value{
an S3 object of class \code{fileAndDataName}
}
\description{
This is an internal \code{\link{dryworkflow-package}} function. It
is primarily designed to be called by
\code{\link{createProjectSkeleton}} and \code{\link{addFile}} to
set up file and object names for processing. Given a list of data
filenames and optionally, a list of project steps, various names
are created for use with templates, makefiles and git for version
control.
}

