% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{whichReadCommand}
\alias{whichReadCommand}
\title{Provide appropriate read commands for reading data files}
\usage{
whichReadCommand(data.files, commands = list(tsv = "read.delim", csv =
  "read.csv", dta = "foreign::read.dta", sas7bdat = "sas7bdat::read.sas7bdat",
  xport = "foreign::read.xport", sav = "foreign::spss.get", rdata = "load", rda
  = "load", xls = "readxl::read_excel", xlsx = "readxl::read_excel"),
  other = NULL, extra.args = list(tsv = NA, csv =
  "na.strings = c(\\"\\", \\"NA\\", \\".\\")", dta = NA, sas7bdat = NA, xport =
  NA, sav = NA, rdata = NA, rda = NA, xls = NA, xlsx = NA),
  unsupported = "UNSUPPORTED:PUT_READ_FUNCTION_HERE")
}
\arguments{
\item{data.files}{character vector of file names}

\item{commands}{list of strings to be substituted into R syntax
files to read each file in \code{data.files}}

\item{other}{list of either alternate or extra commands for
reading data files}

\item{extra.args}{extra arguments that will be be added to the
read command produced}

\item{unsupported}{string to use if filetype not supported}
}
\value{
named vector of characters to substitute for read commands
in \code{R} syntax file. The \dQuote{extraLibs} attribute is a
character vector of library commands which can be incorporated
into syntax files although this is unnecessary because functions
are called directly
}
\description{
Using the filename extensions, such as \code{.csv} or \code{.dta},
a best match is determined for inclusion in \code{R} syntax files
for reading data. Suggested commands can be changed or extras
added by using the \code{other} option.
}
\author{
Peter Baker
}
\keyword{internal}

